requirements.txt
phidata
fpdf
python-dotenv
rich
typer
requests
numpy
qdrant-client
groq
fastembed
pypdf